# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Food API Documentation",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Food API"
        }
    ],
    "paths": {
...
        "/intakes": {
            "get": {
                "tags": [
                    "Intakes"
                ],
                "summary": "Get all intakes",
                "description": "Get all intakes",
                "operationId": "2fd2d876107aabea55cd8837f2a3ecf4",
                "responses": {
                    "200": {
                        "description": "All intakes"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Intakes"
                ],
                "summary": "Create an intake",
                "description": "Create an intake",
                "operationId": "0935d69845b8db6f05f8af1fffecc3d3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "product_id": {
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "default": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Intake created successfully"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/intakes/{id}": {
            "get": {
                "tags": [
                    "Intakes"
                ],
                "summary": "Get an intake by ID",
                "description": "Get an intake by ID",
                "operationId": "833d923f86839c35b77c600cc329df28",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Intake ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Intake"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Intakes"
                ],
                "summary": "Update an intake",
                "description": "Update an intake",
                "operationId": "ac8036738f33eda4ccb6e2c90621a75c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Intake ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "product_id": {
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "default": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Intake updated successfully"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Intakes"
                ],
                "summary": "Delete an intake",
                "description": "Delete an intake",
                "operationId": "e139088a773f74dbed12edbe54c26a6e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Intake ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Intake deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/off/food/search/{query}/{page}": {
            "get": {
                "tags": [
                    "Open Food Facts"
                ],
                "summary": "Search for food",
                "operationId": "f75a26492983967dc95631e9e14cad35",
                "parameters": [
                    {
                        "name": "query",
                        "in": "path",
                        "description": "Search query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of food products from Open Food Facts"
                    }
                }
            }
        },
        "/off/food/{id}": {
            "get": {
                "tags": [
                    "Open Food Facts"
                ],
                "summary": "Get food product",
                "operationId": "765edf11ba09ff124f48924edd2e288b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}